Major update
This update contains many improvements of plots and text. The code is cleaned-up and updated to the current version of the Spotify API and the R package Spotifyr.

> Files

Added project file 'bowie.Rproj'.

Added file 'spotify.R' for personal API credentials.

New file 'bowie.R' contains functions that were previously in 'index.Rmd', making the code more readable.

Included R libraries are now in a folder called 'index_files'. This reduces the file size of 'index.html' from 10.5 MB to 574 KB.

> Custom scatterplots

Changed legend background to the same color as the plot, to improve legibility.

Removed plot of sadness versus instrumentalness. Sadness was computed as the distance to the sad corner of the valence-energy plot. This approach did not offer substantial new insights.

Improved tooltips (information boxes that appear when the cursor is positioned over an element):
- Removed 'danceability' score from track tooltips, because it is not visualized in the plots, and leads to an excess of information.
- Removed unintended tooltips of the four emotions in the Emotion chart.
- Removed tooltip 'trilogy within'. The text aesthetic was set in geom_polygon(), resulting in an 'ignoring unknown aesthetics: text' warning from the compiler. The triangle is now drawn using geom_path().